Fix Unity Cloud Build issue #172
Prior to this change AndroidSdkPlatform.cs would check to see if an Android SDK with version 30 and above was installed and throw an error if it was missing.

Now AndroidSdkPlatform.cs only checks the installed Android SDK if targetAndroidSdk is set to "auto (highest installed)", or a version less than 30.

See issue #172 for context.

PiperOrigin-RevId: 432506826